Move single-use util functions into client package
