+++ pkg/utils/user_agent.go
@@ -9,8 +9,8 @@ const FirefoxLinux = "Mozilla/5.0 (X11; Linux x86_64; rv:88.0) Gecko/20100101 Fi
 const FirefoxLinuxArm = "Mozilla/5.0 (X11; Linux armv7l; rv:86.0) Gecko/20100101 Firefox/86.0"
 const FirefoxLinuxArm64 = "Mozilla/5.0 (X11; Linux aarch64; rv:86.0) Gecko/20100101 Firefox/86.0"
 
-// GetUserAgent returns a valid User Agent string that matches the running os/arch
-func GetUserAgent() string {
+// getUserAgent returns a valid User Agent string that matches the running os/arch
+func getUserAgent() string {
 	arch := runtime.GOARCH
 	os := runtime.GOOS
